[TASK] register elements in AfterTcaCompilationEvent
with the "wizardItemsHook"-hook you can only manipulate the final wizard-items array.
This has several drawbacks, like you have to apply translation of labels manually.
But the real show-stopper  is the fact that the NewContentElementController skips empty wizard groups.
Once you we use such a removed group that won't exist anymore inside the generated array, the CE "moves" to the next available wizard-group.
That is not desirable and hard to avoid.